jobagent/requirements.txt
streamlit
requests
beautifulsoup4
lxml
python-dateutil
PyYAML
apscheduler
pandas
scikit-learn
sentence-transformers
PyPDF2
python-docx
streamlit
sentence-transformers
beautifulsoup4
lxml
requests
PyPDF2
python-docx
apscheduler
pandas
pyyaml
